cleaned up commander
