+++ app/src/main/java/com/rx/demo/commander/UserCommander.java
@@ -12,7 +12,9 @@ import javax.inject.Singleton;
  * Created by Nakhimovich on 3/25/15.
  */
 @Singleton
-public class UserCommander extends Commander<Object, ArrayList<User>> {
+public class UserCommander extends Commander<ArrayList<User>> {
+    private final String accessToken="6b060e39a5cf06b980ec4f1d2042b525026c35d2";
+
     @Inject
     Github api;
 
@@ -22,6 +24,6 @@ public class UserCommander extends Commander<Object, ArrayList<User>> {
 
     @Override
     public ArrayList<User> load(Object request) throws Exception {
-        return api.users();
+        return api.users("token " + accessToken);
     }
 }
