+++ app/src/main/java/com/rx/demo/rest/Github.java
@@ -5,10 +5,12 @@ import com.rx.demo.model.User;
 import java.util.ArrayList;
 
 import retrofit.http.GET;
+import retrofit.http.Header;
+
 /**
  * Created by MikeN on 8/16/14.
  */
 public interface Github {
     @GET("/users")
-    ArrayList<User> users();
+    ArrayList<User> users(@Header("Authorization")String accessToken);
 }
